Adds embed mode to CodeViewer and DiffViewer
Adds an `isEmbed` prop to CodeViewer and DiffViewer components.

When `isEmbed` is true, the component's box-shadow and border-radius are removed, and overflow is set to auto.
This is useful for embedding the component in other contexts where a more seamless integration is desired.

Also, now `useCodeDrawer` and `useDiffDrawer` accept `lang`, `originalLang`, and `modifiedLang` to define code language.